Change status parameter type to Short in getAppointmentsByStatus method

+++ src/main/kotlin/com/icare/service/AppointmentService.kt
@@ -4,10 +4,9 @@ import com.icare.model.Appointment
 import com.icare.model.ResponseModel
 
 
-
 interface AppointmentService {
-    fun bookAppointment(Appointments: Appointment ): ResponseModel
-    fun  getPatientAppointments(token: String): ResponseModel
-    fun getAppointmentsByStatus(status: Int,token: String): ResponseModel
+    fun bookAppointment(Appointments: Appointment): ResponseModel
+    fun getPatientAppointments(token: String): ResponseModel
+    fun getAppointmentsByStatus(status: Short, token: String): ResponseModel
 
 }
